+++ CHANGELOG.md
@@ -1,7 +1,9 @@
 # Changelog
-### Unreleased
+### KiteUtils v0.9.6 - 2024-12-19
 #### Added
 - function `KiteUtils.install_examples()`
+- add field `upwind_dir`, remove vector `v_wind_ref` from `Settings`and yaml files
+- add the fields `max_acc`, `p_speed` and `i_speed` to `Settings`and yaml files
 
 ### KiteUtils v0.9.5 - 2024-12-06
 #### Changed